Added threading to queue jobs.
The server will now operate two threads for handling jobs. Need to check
and make sure performance is stable across all modules but improved
speed of getting artist details in the queue. Tested pausing and
resuming queue for scanner to do it's thing. Scanner might be running
too fast, a way to fix that would be to double the individual iteration
time to allow for both threads to continue running.