Implemented unit test/method for staging dummy database.
Implemented a staging method for use with unit tests that will help with
testing the email new release service. It creates two users and four
artists, each user having a different email address and each artist
having either a new release or a standard up to date release. The unit
test confirms that the data is serialized before resolving.