Implemented send new release emails for handler.
Implemented the method for sending out new releases once the new release
handler finishes processing all of the artists. The unit test for the
new release handler works for the email handler, but still need to
duplicate the code over to the other unit test file and change the
assertions.