Implemented bug fix for missing release dates, implemented unit tests to check new functionality.
Added a call to the get artists details queue when a new release has
been found for an artist. We were running into issues with release dates
not being added to artists when releases are found. Now when new
releases are added, the artist document will be updated with the date.
Slightly modified the parameters for getArtistDetails so that when you
pass an artist object you don't have to create an object in the
parameter call. So instead of {artist: artist} you can just pass artist.
Added a logic gate to check if the socketUtil exists in this instance of
the artist details queue module to avoid unnecessary error logs while
running unit tests. Removed a few unnecessary console logs in the
test-helpers module. Slightly modified unit tests variable declarations
for better code readability.