Made extensive changes to how new releases are parsed.
Changed how new releases are parsed in the new releases utility by
clearing special characters to avoid Spotify typos with symbols, and
checked to make sure that the releases that are obtained are actually
the most recent. The tag:new query for spotify does not include releases
that came out today so I made sure to double check while parsing the
release array just in case the service finds out a new release before
spotify propogates it. There is still a lot of work to be done to
eliminate code duplication, but it's stable right now. I modified some
test values to represent data structure modifications, particularly how
the releases are returned from their function call. Now it is an object
with the spotify id of the artist being the key and the releases being
an array tied to the key.