Continue implementation of playlist-handler helpers
Remove unnecessary documentation of user schema.

Create two new spotify-server-api methods, getAlbumTracks, which
promise-wraps the getAlbumtracks api call and refreshes server client
token if necessary, and getAlbumTracksItems, which returns a subset of
the priors promise return.

Delete some unnecassary deprecated commented code in spotify-user-api.

Create and modified helper functions in spotify-user-api in order to
retrieve uris for tracks to add to playlist for a user.